refactor: 패키지명에서 언더바 제거
