+++ core/src/main/kotlin/team/duckie/quackquack/core/animation/QuackAnimatedVisibility.kt
@@ -15,7 +15,7 @@ import androidx.compose.animation.fadeIn
 import androidx.compose.animation.fadeOut
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
-import team.duckie.quackquack.core._material.QuackAnimationSpec
+import team.duckie.quackquack.core.material.QuackAnimationSpec
 
 /**
  * 컴포저블의 visiblilty 변화에 애니메이션을 적용하는 컨테이너 입니다.
