+++ core/src/main/kotlin/team/duckie/quackquack/core/component/text.kt
@@ -25,12 +25,12 @@ import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.util.fastForEach
 import androidx.compose.ui.util.fastMap
 import kotlinx.collections.immutable.ImmutableList
-import team.duckie.quackquack.core._internal.QuackDataModifierModel
-import team.duckie.quackquack.core._internal.quackComposed
-import team.duckie.quackquack.core._internal.quackMaterializerOf
-import team.duckie.quackquack.core._material.QuackColor
-import team.duckie.quackquack.core._material.QuackTypography
-import team.duckie.quackquack.core._material.animatedQuackTextStyleAsState
+import team.duckie.quackquack.core.runtime.QuackDataModifierModel
+import team.duckie.quackquack.core.runtime.quackComposed
+import team.duckie.quackquack.core.runtime.quackMaterializerOf
+import team.duckie.quackquack.core.material.QuackColor
+import team.duckie.quackquack.core.material.QuackTypography
+import team.duckie.quackquack.core.material.animatedQuackTextStyleAsState
 import team.duckie.quackquack.core.animation.QuackAnimatedContent
 import team.duckie.quackquack.core.sugar.annotation.Sugar
 import team.duckie.quackquack.core.util.fastFirstInstanceOrNull
