+++ core/src/main/kotlin/team/duckie/quackquack/core/runtime/model.kt
@@ -5,7 +5,7 @@
  * Please see full license: https://github.com/duckie-team/duckie-quack-quack/blob/main/LICENSE
  */
 
-package team.duckie.quackquack.core._internal
+package team.duckie.quackquack.core.runtime
 
 import androidx.compose.runtime.Immutable
 import androidx.compose.ui.Modifier
