+++ core/src/main/kotlin/team/duckie/quackquack/core/theme/theme.kt
@@ -15,7 +15,7 @@ import androidx.compose.runtime.Composable
 import androidx.compose.runtime.CompositionLocalProvider
 import androidx.compose.runtime.ProvidableCompositionLocal
 import androidx.compose.runtime.staticCompositionLocalOf
-import team.duckie.quackquack.core._material.QuackColor
+import team.duckie.quackquack.core.material.QuackColor
 
 /**
  * QuackTextField 및 QuackTextArea에서 사용할 커서 색상을 제공합니다.
